src/kafka_producer/requirements.txt
kafka-python==2.0.2
mysql-connector-python
requests==2.26.0


